Ported back to WIN32
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@9745 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/src/c/socket/xmlBlasterSocket.h
@@ -64,7 +64,7 @@ enum XMLBLASTER_MSG_TYPE_ENUM {
 #define XMLBLASTER_VERSION 49
 
 
-extern int readn(int fd, char *ptr, int nbytes);
+extern size_t readn(int fd, char *ptr, size_t nbytes);
 extern int getLength(char *data);
 extern int isListening();
 extern void shutdownCallbackServer();
